# Optimize conversion to unstructured