--- a/value/mapreflect.go
+++ b/value/mapreflect.go
@@ -34,7 +34,7 @@ func (r mapReflect) Get(key string) (Value, bool) {
 	if !ok {
 		return nil, false
 	}
-	return mustWrapValueReflectMapItem(&r.Value, &k, v), true
+	return mustWrapValueReflect(v, &r.Value, &k), true
 }
 
 func (r mapReflect) get(k string) (key, value reflect.Value, ok bool) {
@@ -70,19 +70,19 @@ func (r mapReflect) toMapKey(key string) reflect.Value {
 func (r mapReflect) Iterate(fn func(string, Value) bool) bool {
 	vr := reflectPool.Get().(*valueReflect)
 	defer vr.Recycle()
-	return eachMapEntry(r.Value, func(s string, value reflect.Value) bool {
-		return fn(s, vr.reuse(value))
+	return eachMapEntry(r.Value, func(key reflect.Value, value reflect.Value) bool {
+		return fn(key.String(), vr.mustReuse(value, &r.Value, &key))
 	})
 }
 
-func eachMapEntry(val reflect.Value, fn func(string, reflect.Value) bool) bool {
+func eachMapEntry(val reflect.Value, fn func(reflect.Value, reflect.Value) bool) bool {
 	iter := val.MapRange()
 	for iter.Next() {
 		next := iter.Value()
 		if !next.IsValid() {
 			continue
 		}
-		if !fn(iter.Key().String(), next) {
+		if !fn(iter.Key(), next) {
 			return false
 		}
 	}
@@ -114,6 +114,6 @@ func (r mapReflect) Equals(m Map) bool {
 		if !ok {
 			return false
 		}
-		return Equals(vr.reuse(lhsVal), value)
+		return Equals(vr.mustReuse(lhsVal, nil, nil), value)
 	})
 }